src/services/auth/lambda/requirements.txt
PyJWT==2.8.0
boto3==1.34.0
botocore==1.34.0 